requirement.txt
scipy
scikit-learn
numpy
opencv_python
Pillow
torch>=1.2.0
torchvision
pyyaml
tqdm
tensorboardX
ftfy
regex
einops
timm
pandas
matplotlib
transformers==4.47.0
torchmetrics==1.5.0
diffusers==0.29.2
autokeras==1.0.20
tensorflow-text==2.8.2
tensorflow
huggingface_hub==0.25.2
git+https://github.com/ml-research/safe-latent-diffusion.git
tabulate
pytorch-fid
sentencepiece
protobuf==3.20.0
nudenet>=3.4.2
kornia>=0.8.0 
kornia_rs>=0.1.8
open-clip-torch-2.22.0
rich
ml_collections
lpips